% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besd_ch_core.R
\name{besd_ch_core}
\alias{besd_ch_core}
\title{Calculate core BeSD indicators for a child vaccination analysis}
\usage{
besd_ch_core(VCP = "besd_ch_core", cleanup = TRUE)
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}

\item{cleanup}{Clean up global settings after run}
}
\value{
Log entries; errors if conditions not met
}
\description{
Calculate core BeSD indicators for a child vaccination analysis
}
\examples{
besd_ch_core()
}
