% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besd_copy.R
\name{besd_copy_multilingual}
\alias{besd_copy_multilingual}
\title{Copy template files to a target folder}
\usage{
besd_copy_multilingual(folder, overwrite = TRUE)
}
\arguments{
\item{folder}{Folder where file should be copied}

\item{overwrite}{Overwrite existing destination file?}
}
\value{
File copied to target folder
}
\description{
Copy template files to a target folder
}
