% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besd_core_plot.R
\name{besd_core_plot}
\alias{besd_core_plot}
\title{Make core indicator barplot}
\usage{
besd_core_plot(VCP = "besd_core_plot", analysis)
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}

\item{analysis}{Type of BeSD-TI analysis ("child", "covid", etc.)}
}
\value{
Core indicator barplot
}
\description{
Make core indicator barplot
}
