% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besd_covid_core.R
\name{besd_covid_core}
\alias{besd_covid_core}
\title{Calculate core BeSD indicators for a COVID-19 vaccination analysis}
\usage{
besd_covid_core(VCP = "besd_covid_core", cleanup = TRUE)
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}
}
\value{
Log entries; errors if conditions not met
}
\description{
Calculate core BeSD indicators for a COVID-19 vaccination analysis
}
\examples{
besd_covid_core()
}
