% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BESD_DESC_02.R
\name{BESD_DESC_02}
\alias{BESD_DESC_02}
\title{Response to multiple-choice question with only one response}
\usage{
BESD_DESC_02(VCP = "BESD_DESC_02", database_id = NA, cleanup = FALSE)
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}

\item{database_id}{Custom tag for output file names (optional)}

\item{cleanup}{Clean up DESC_02 global settings after run}
}
\value{
Derived variables, databases and tables
}
\description{
Response to multiple-choice question with only one response
}
\examples{
BESD_DESC_02()
}
