% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BESD_DESC_02_00GC.R
\name{BESD_DESC_02_00GC}
\alias{BESD_DESC_02_00GC}
\title{Check global macros for BESD_DESC_02}
\usage{
BESD_DESC_02_00GC(VCP = "BESD_DESC_02_00GC")
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}
}
\value{
Errors if conditions not met
}
\description{
Check global macros for BESD_DESC_02
}
