% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BESD_DESC_02_01PP.R
\name{BESD_DESC_02_01PP}
\alias{BESD_DESC_02_01PP}
\title{Pre-process dataset for BESD_DESC_02}
\usage{
BESD_DESC_02_01PP(VCP = "BESD_DESC_02_01PP", database_id = NA)
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}
}
\value{
A dataset (BESD_DESC_02_<ANALYSIS_COUNTER>_<BESD_DESC_02_COUNTER>)
}
\description{
Pre-process dataset for BESD_DESC_02
}
