% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BESD_DESC_02_03DV.R
\name{BESD_DESC_02_03DV}
\alias{BESD_DESC_02_03DV}
\title{Calculate derived variables for BESD_DESC_02}
\usage{
BESD_DESC_02_03DV(VCP = "BESD_DESC_02_03DV", database_id = NA)
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}

\item{database_id}{Custom tag for output file names (optional)}
}
\value{
A dataset (BESD_DESC_02_<ANALYSIS_COUNTER>_<BESD_DESC_02_COUNTER>)
}
\description{
Calculate derived variables for BESD_DESC_02
}
