% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BESD_DESC_02_07RT.R
\name{BESD_DESC_02_07RT}
\alias{BESD_DESC_02_07RT}
\title{Create report components for BESD_DESC_02}
\usage{
BESD_DESC_02_07RT(VCP = "BESD_DESC_02_07RT", database_id = NA)
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}

\item{database_id}{Tag for file names}
}
\value{
Table to be inserted in template report
}
\description{
Create report components for BESD_DESC_02
}
