% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BESD_DESC_03.R
\name{BESD_DESC_03}
\alias{BESD_DESC_03}
\title{Response to multiple-choice question where the respondent may select more than one response option.}
\usage{
BESD_DESC_03(VCP = "BESD_DESC_03", database_id = NA, cleanup = FALSE)
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}

\item{database_id}{Tag for file names}

\item{cleanup}{Clean up DESC_03 global settings after run}
}
\value{
Derived variables, databases and tables
}
\description{
Response to multiple-choice question where the respondent may select more than one response option.
}
\examples{
BESD_DESC_03()
}
