% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BESD_DESC_03_01PP.R
\name{BESD_DESC_03_01PP}
\alias{BESD_DESC_03_01PP}
\title{Pre-process dataset for DESC_03}
\usage{
BESD_DESC_03_01PP(VCP = "BESD_DESC_03_01PP", database_id = NA)
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}

\item{database_id}{Custom tag for output file names (optional)}
}
\value{
A dataset (DESC_03_<ANALYSIS_COUNTER>_<DESC_03_COUNTER>)
}
\description{
Pre-process dataset for DESC_03
}
