% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BESD_DESC_03_03DV.R
\name{BESD_DESC_03_03DV}
\alias{BESD_DESC_03_03DV}
\title{Calculate derived variables for DESC_03}
\usage{
BESD_DESC_03_03DV(VCP = "BESD_DESC_03_03DV", database_id = NA)
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}

\item{database_id}{Custom tag for output file names (optional)}
}
\value{
A dataset (DESC_03_<ANALYSIS_COUNTER>_<DESC_03_COUNTER>)
}
\description{
Calculate derived variables for DESC_03
}
