% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BESD_DESC_03_04GO.R
\name{BESD_DESC_03_04GO}
\alias{BESD_DESC_03_04GO}
\title{Generate output databases for DESC_03}
\usage{
BESD_DESC_03_04GO(VCP = "BESD_DESC_03_04GO", database_id = NA)
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}

\item{database_id}{Tag for file names}
}
\value{
Databases in OUTPUT_FOLDER
}
\description{
Generate output databases for DESC_03
}
