% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BESD_DESC_03_05TOST.R
\name{BESD_DESC_03_05TOST}
\alias{BESD_DESC_03_05TOST}
\title{Export datasets to Excel for DESC_03}
\usage{
BESD_DESC_03_05TOST(VCP = "BESD_DESC_03_05TOST", database_id = NA)
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}

\item{database_id}{Custom tag for output file names (optional)}
}
\value{
Sheet(s) in tabular output Excel file in OUTPUT_FOLDER
}
\description{
Export datasets to Excel for DESC_03
}
