% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besd_global.R
\name{besd_global}
\alias{besd_global}
\title{Define and log a value in the global environment}
\usage{
besd_global(global, value)
}
\arguments{
\item{global}{The name of the global value}

\item{value}{The value to be assigned}
}
\value{
An R object in the global environment
}
\description{
Define and log a value in the global environment
}
\examples{
besd_global(x, 100)
besd_global(y, c("a", "b", "c"))

}
