% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besd_halt_immediately.R
\name{besd_halt_immediately}
\alias{besd_halt_immediately}
\title{Exit BeSD-TI program at the end of a run or after an error}
\usage{
besd_halt_immediately(VCP = "besd_halt_immediately", halt_message = NA)
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}

\item{halt_message}{Error text to print when BeSD-TI is stopped mid-run}
}
\value{
Clean exit from BeSD-TI programs
}
\description{
Exit BeSD-TI program at the end of a run or after an error
}
