% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besd_log_global.R
\name{besd_log_global}
\alias{besd_log_global}
\title{Put the current value of a global into the BeSD-TI log}
\usage{
besd_log_global(global)
}
\arguments{
\item{global}{A global value to be logged}
}
\value{
Call to besd_log_comment
}
\description{
Put the current value of a global into the BeSD-TI log
}
\examples{
test_value <- 1
besd_log_global(test_value)
}
