% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besd_multi_lingual_strings.R
\name{besd_multi_lingual_strings}
\alias{besd_multi_lingual_strings}
\title{Check language choice and language file for BeSD}
\usage{
besd_multi_lingual_strings(VCP = "besd_multi_lingual_strings")
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}
}
\value{
A global that records the language choice and errors and/or warnings if conditions not met
}
\description{
Check language choice and language file for BeSD
}
\examples{
besd_multi_lingual_strings()
}
