% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besd_read.R
\name{besd_read}
\alias{besd_read}
\title{Detect file type and read dataset}
\usage{
besd_read(file)
}
\arguments{
\item{file}{File to read}
}
\value{
A data frame or tibble
}
\description{
Detect file type and read dataset
}
