% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besd_report.R
\name{besd_report}
\alias{besd_report}
\title{Make template BeSD survey report}
\usage{
besd_report(VCP = "besd_report", filetype = "docx", analysis)
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}

\item{filetype}{Report output format}

\item{analysis}{Type of BeSD-TI analysis ("child", "covid", etc.)}
}
\value{
Template BeSD survey report
}
\description{
Make template BeSD survey report
}
\examples{
besd_report(analysis = "child")
}
