% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besd_report_setup.R
\name{besd_report_setup}
\alias{besd_report_setup}
\title{Copy report template to output folder}
\usage{
besd_report_setup(template_path = NULL, outfile_path = NULL)
}
\arguments{
\item{template_path}{Path to template to copy}

\item{outfile_path}{Path where copy of template should be saved (usually in OUTPUT_FOLDER)}
}
\value{
Template files copied to output folder
}
\description{
Copy report template to output folder
}
