% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_besd_analysis_metadata.R
\name{check_besd_analysis_metadata}
\alias{check_besd_analysis_metadata}
\title{Check BeSD-TI analysis metadata for missing information and format problems}
\usage{
check_besd_analysis_metadata(VCP = "check_besd_analysis_metadata")
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}
}
\value{
Default run settings, default level4 layout file if needed, errors and/or warnings if conditions not met
}
\description{
Check BeSD-TI analysis metadata for missing information and format problems
}
\examples{
check_besd_analysis_metadata()

}
