% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_besd_metadata.R
\name{check_besd_metadata}
\alias{check_besd_metadata}
\title{Check BeSD-TI analysis-related globals, datasets and variables}
\usage{
check_besd_metadata(VCP = "check_besd_metadata", analysis)
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Log entries; errors if conditions not met
}
\description{
Check BeSD-TI analysis-related globals, datasets and variables
}
\examples{
check_besd_metadata()
}
