% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_BeSDTI_globals.R
\name{cleanup_BeSDTI_globals}
\alias{cleanup_BeSDTI_globals}
\title{Clean up BeSD-TI globals}
\usage{
cleanup_BeSDTI_globals()
}
\value{
Remove BeSD-TI global values from global environment
}
\description{
Clean up BeSD-TI globals
}
