% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_unique_CH_ids.R
\name{establish_unique_CH_ids}
\alias{establish_unique_CH_ids}
\title{Check/create unique ID variables in CH dataset and merge stratifiers}
\usage{
establish_unique_CH_ids(VCP = "establish_unique_CH_ids")
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}
}
\value{
CH_with_ids dataset in OUTPUT_FOLDER
}
\description{
Check/create unique ID variables in CH dataset and merge stratifiers
}
\examples{
establish_unique_CH_ids()
}
