% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_unique_COV_ids.R
\name{establish_unique_COV_ids}
\alias{establish_unique_COV_ids}
\title{Check/create unique ID variables in COVID-19 dataset and merge stratifiers}
\usage{
establish_unique_COV_ids(VCP = "establish_unique_COV_ids")
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}
}
\value{
COV_with_ids dataset in OUTPUT_FOLDER
}
\description{
Check/create unique ID variables in COVID-19 dataset and merge stratifiers
}
\examples{
establish_unique_COV_ids()
}
