% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ch_dv.R
\name{gen_ch_dv}
\alias{gen_ch_dv}
\title{Reads in the child vaccination BeSD survey and creates the derived variables}
\usage{
gen_ch_dv(VCP = "gen_ch_dv")
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}
}
\value{
a dataset
}
\description{
Reads in the child vaccination BeSD survey and creates the derived variables
}
\examples{
gen_ch_dv()
}
