% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cov_dv.R
\name{gen_cov_dv}
\alias{gen_cov_dv}
\title{Reads in the COVID-19 BeSD survey and creates the derived variables}
\usage{
gen_cov_dv(VCP = "gen_cov_dv")
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}
}
\value{
a dataset
}
\description{
Reads in the COVID-19 BeSD survey and creates the derived variables
}
\examples{
gen_cov_dv()
}
