% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_besd_output_database.R
\name{make_besd_output_database}
\alias{make_besd_output_database}
\title{Create database for BeSD core indicators}
\usage{
make_besd_output_database(
  variable,
  label = NA,
  weighted,
  denominator,
  VCP = "make_besd_output_database",
  ...
)
}
\arguments{
\item{variable}{Outcome variable}

\item{label}{Label for the outcome variable}

\item{weighted}{Calculate weighted indicator (TRUE) or unweighted (FALSE)}

\item{denominator}{Which respondents to include in the denominator}

\item{VCP}{Current program name to be logged, default to be the function name}

\item{...}{Other arguments}
}
\value{
A database in OUTPUT_FOLDER
}
\description{
Create database for BeSD core indicators
}
