% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DESC_0203_output_database.R
\name{make_DESC_0203_output_database}
\alias{make_DESC_0203_output_database}
\title{Create database for DESC_02 and DESC_03}
\usage{
make_DESC_0203_output_database(
  variable,
  label = NA,
  vid,
  measureid,
  printprogress = NULL,
  VCP = "make_DESC_0203_output_database",
  database_id = NA,
  ...
)
}
\arguments{
\item{variable}{Outcome variable}

\item{label}{Label for the outcome variable}

\item{vid}{Outcome variable ID for filenames}

\item{measureid}{Analysis indicator ID}

\item{printprogress}{Progress tracking string to print}

\item{VCP}{Current program name to be logged, default to be the function name}

\item{database_id}{Custom tag for output file names (optional)}

\item{...}{Other arguments}
}
\value{
A database in OUTPUT_FOLDER
}
\description{
Create database for DESC_02 and DESC_03
}
