% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besd_core_plot.R
\docType{data}
\name{xlab_wrap}
\alias{xlab_wrap}
\title{Make core indicator barplot}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
xlab_wrap
}
\arguments{
\item{VCP}{Current program name to be logged, default to be the function name}

\item{analysis}{Type of BeSD-TI analysis ("child", "covid", etc.)}
}
\value{
Core indicator barplot
}
\description{
Make core indicator barplot
}
\keyword{datasets}
