% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besd_report_setup.R
\name{besd_report_setup}
\alias{besd_report_setup}
\title{Copy report template to output folder}
\usage{
besd_report_setup(template_name = NULL, outfile_name = NULL)
}
\arguments{
\item{template_name}{Name of template to copy}

\item{outfile_name}{Name for copy of template}
}
\value{
Template files copied to output folder
}
\description{
Copy report template to output folder
}
